refactor(packages/rspack): remove `Rspack` from server/index.ts to index.ts
