+++ packages/rspack/src/build.ts
@@ -1,4 +1,4 @@
-import { Rspack } from './server';
+import { Rspack } from '.';
 export async function build(config: any) {
   const rspack = new Rspack(config);
   return await rspack.build();
